Refactors code structure according to clean architecture
