+++ lib/src/data/repositories/genre_repository.dart
@@ -2,8 +2,8 @@ import 'dart:convert';
 
 import 'package:flutter/services.dart';
 
-import '../interfaces/i_repository.dart';
-import '../models/genres_model.dart';
+import '../../domain/repository/i_repository.dart';
+import '../../domain/entity/genres_model.dart';
 
 class GenreRepository implements IRepository<GenresModel> {
   @override
