+++ lib/src/domain/entity/movie.dart
