+++ lib/src/domain/entity/movies_model.dart
