+++ pubspec.yaml
@@ -45,6 +45,7 @@ dev_dependencies:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
+  lint: ^2.0.0
   flutter_lints: ^2.0.0
 
 # For information on the generic Dart part of this file, see the
